requirements.txt
psutil
flask
py-cpuinfo